# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": ["code", "message", "details"],
        "title": "ApiError",
        "type": "object"
      },
...
      "DatasetStatsModel": {
        "additionalProperties": false,
        "properties": {
          "modules": {
            "items": {
              "$ref": "#/components/schemas/ModuleStatsModel"
            },
            "title": "Modules",
            "type": "array"
          },
          "named_graphs": {
            "title": "Named Graphs",
            "type": "integer"
          },
          "quads": {
            "title": "Quads",
            "type": "integer"
          },
          "resources": {
            "title": "Resources",
            "type": "integer"
          },
          "types": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Types",
            "type": "object"
          }
        },
        "required": ["quads", "named_graphs", "resources", "types", "modules"],
        "title": "DatasetStatsModel",
        "type": "object"
      },
...
      "ModuleGraphModel": {
        "additionalProperties": false,
        "properties": {
          "graph_iri": {
            "title": "Graph Iri",
            "type": "string"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "module_id": {
            "title": "Module Id",
            "type": "string"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "quads": {
            "items": {
              "$ref": "#/components/schemas/QuadModel"
            },
            "title": "Quads",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "module_id",
          "graph_iri",
          "total",
          "offset",
          "limit",
          "truncated",
          "quads"
        ],
        "title": "ModuleGraphModel",
        "type": "object"
      },
      "ModuleModel": {
        "additionalProperties": false,
        "properties": {
          "classes": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Classes",
            "type": "array"
          },
          "competency_question_count": {
            "title": "Competency Question Count",
            "type": "integer"
          },
          "definitions": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Definitions",
            "type": "array"
          },
          "graph_iri": {
            "title": "Graph Iri",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "import_cycles": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Import Cycles",
            "type": "array"
          },
          "imports": {
            "items": {
              "type": "string"
            },
            "title": "Imports",
            "type": "array"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Labels",
            "type": "array"
          },
          "ontology_iri": {
            "title": "Ontology Iri",
            "type": "string"
          },
          "properties": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Properties",
            "type": "array"
          },
          "responsible": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Responsible",
            "type": "array"
          },
          "source_path": {
            "title": "Source Path",
            "type": "string"
          },
          "status": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Status",
            "type": "array"
          },
          "term_count": {
            "title": "Term Count",
            "type": "integer"
          },
          "term_limit": {
            "title": "Term Limit",
            "type": "integer"
          },
          "term_offset": {
            "title": "Term Offset",
            "type": "integer"
          },
          "term_total": {
            "title": "Term Total",
            "type": "integer"
          },
          "terms_has_next": {
            "title": "Terms Has Next",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "ontology_iri",
          "graph_iri",
          "source_path",
          "imports",
          "labels",
          "definitions",
          "responsible",
          "status",
          "classes",
          "properties",
          "term_total",
          "term_offset",
          "term_limit",
          "terms_has_next",
          "import_cycles",
          "term_count",
          "competency_question_count"
        ],
        "title": "ModuleModel",
        "type": "object"
      },
      "ModulePageModel": {
        "additionalProperties": false,
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ModuleModel"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["items", "total", "offset", "limit", "has_next"],
        "title": "ModulePageModel",
        "type": "object"
      },
      "ModuleStatsModel": {
        "additionalProperties": false,
        "properties": {
          "graph_iri": {
            "title": "Graph Iri",
            "type": "string"
          },
          "module_id": {
            "title": "Module Id",
            "type": "string"
          },
          "quads": {
            "title": "Quads",
            "type": "integer"
          },
          "resources": {
            "title": "Resources",
            "type": "integer"
          },
          "types": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Types",
            "type": "object"
          }
        },
        "required": ["module_id", "graph_iri", "quads", "resources", "types"],
        "title": "ModuleStatsModel",
        "type": "object"
      },
...
      "QuadModel": {
        "additionalProperties": false,
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/RdfValueModel"
          },
          "object": {
            "$ref": "#/components/schemas/RdfValueModel"
          },
          "predicate": {
            "$ref": "#/components/schemas/RdfValueModel"
          },
          "priority": {
            "default": 7,
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Priority",
            "type": "integer"
          },
          "relationship_kind": {
            "default": "other",
            "enum": [
              "subclass",
              "subproperty",
              "domain",
              "range",
              "import",
              "internal_object_property",
              "other"
            ],
            "title": "Relationship Kind",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "subject": {
            "$ref": "#/components/schemas/RdfValueModel"
          }
        },
        "required": ["subject", "predicate", "object", "graph"],
        "title": "QuadModel",
        "type": "object"
      },
      "RdfValueModel": {
        "additionalProperties": false,
        "properties": {
          "compact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compact"
          },
          "datatype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datatype"
          },
          "kind": {
            "enum": ["iri", "bnode", "literal"],
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": ["kind", "value"],
        "title": "RdfValueModel",
        "type": "object"
      },
...
      "RepositoryRevisionModel": {
        "additionalProperties": false,
        "properties": {
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit"
          },
          "dirty": {
            "title": "Dirty",
            "type": "boolean"
          }
        },
        "required": ["branch", "commit", "dirty"],
        "title": "RepositoryRevisionModel",
        "type": "object"
      },
...
      "RuntimeStatusModel": {
        "additionalProperties": false,
        "properties": {
          "generation": {
            "title": "Generation",
            "type": "integer"
          },
          "loaded_at": {
            "title": "Loaded At",
            "type": "string"
          },
          "modules": {
            "title": "Modules",
            "type": "integer"
          },
          "quads": {
            "title": "Quads",
            "type": "integer"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "revision": {
            "$ref": "#/components/schemas/RepositoryRevisionModel"
          },
          "validation_conforms": {
            "title": "Validation Conforms",
            "type": "boolean"
          }
        },
        "required": [
          "ready",
          "generation",
          "loaded_at",
          "revision",
          "quads",
          "modules",
          "validation_conforms"
        ],
        "title": "RuntimeStatusModel",
        "type": "object"
      },
...
      "ValidationIssueModel": {
        "additionalProperties": false,
        "properties": {
          "details": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Details",
            "type": "object"
          },
          "graph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "resource": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource"
          },
          "resource_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Type"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source",
          "rule_id",
          "severity",
          "message",
          "resource",
          "resource_type",
          "path",
          "graph",
          "details"
        ],
        "title": "ValidationIssueModel",
        "type": "object"
      },
      "ValidationReportModel": {
        "additionalProperties": false,
        "properties": {
          "conforms": {
            "title": "Conforms",
            "type": "boolean"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssueModel"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "required": ["conforms", "counts", "issues"],
        "title": "ValidationReportModel",
        "type": "object"
      },
      "WorkspaceModel": {
        "additionalProperties": false,
        "properties": {
          "agent_contract_status": {
            "enum": ["synchronized", "stale", "not_available"],
            "title": "Agent Contract Status",
            "type": "string"
          },
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "class_count": {
            "title": "Class Count",
            "type": "integer"
          },
          "commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit"
          },
          "competency_questions": {
            "title": "Competency Questions",
            "type": "integer"
          },
          "concept_count": {
            "title": "Concept Count",
            "type": "integer"
          },
          "individual_count": {
            "title": "Individual Count",
            "type": "integer"
          },
          "module_count": {
            "title": "Module Count",
            "type": "integer"
          },
          "pending_changes": {
            "title": "Pending Changes",
            "type": "boolean"
          },
          "property_count": {
            "title": "Property Count",
            "type": "integer"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeStatusModel"
          },
          "stats": {
            "$ref": "#/components/schemas/DatasetStatsModel"
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationReportModel"
          },
          "validation_conforms": {
            "title": "Validation Conforms",
            "type": "boolean"
          }
        },
        "required": [
          "runtime",
          "stats",
          "validation",
          "competency_questions",
          "agent_contract_status",
          "branch",
          "commit",
          "pending_changes",
          "module_count",
          "class_count",
          "property_count",
          "concept_count",
          "individual_count",
          "validation_conforms"
        ],
        "title": "WorkspaceModel",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Stable API over the canonical RDF/Git workspace, including controlled proposal authoring, semantic review and the synchronized canonical agent contract.",
    "title": "Enterprise Ontology Workbench API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/graph/module": {
      "get": {
        "operationId": "get_graph_module_api_graph_module_get",
        "parameters": [
          {
            "in": "query",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 1500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleGraphModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Get Graph Module",
        "tags": ["workspace", "graph"]
      }
    },
...
    "/api/modules": {
      "get": {
        "operationId": "list_modules_api_modules_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModulePageModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "List Modules",
        "tags": ["workspace"]
      }
    },
    "/api/modules/{module_id}": {
      "get": {
        "operationId": "get_module_api_modules__module_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "term_offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Term Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "term_limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Term Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Get Module",
        "tags": ["workspace"]
      }
    },
    "/api/modules/{module_id}/graph": {
      "get": {
        "operationId": "get_module_graph_api_modules__module_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 1500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleGraphModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Get Module Graph",
        "tags": ["workspace"]
      }
    },
    "/api/modules/{module_id}/raw": {
      "get": {
        "description": "Return a read-only Turtle view generated by the semantic core.",
        "operationId": "get_module_raw_api_modules__module_id__raw_get",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/turtle": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Read-only Turtle serialization of the module graph."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Get Module Raw",
        "tags": ["workspace"]
      }
    },
...
    "/api/workspace": {
      "get": {
        "operationId": "get_workspace_api_workspace_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Get Workspace",
        "tags": ["workspace"]
      }
    },
    "/api/workspace/reload": {
      "post": {
        "operationId": "reload_workspace_api_workspace_reload_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeStatusModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Reload Workspace",
        "tags": ["workspace"]
      }
    },
...
  }
}